college-marketplace/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3
Pillow==9.5.0
Werkzeug==2.0.1
python-dotenv==0.19.0
